separator evaluator
